Full flow working
Got the full flow of two-way communication working. See
examples/test.py to see example. Uses ZigBee object from the
python-xbee library for the underlying DigiMesh (DM) firmware flashed
on the XBees (S1 and S1-PRO). See notes in send function of Device.py.
Still needs work in the _gotFrame() function used for asynchronous data
reception to allow better parsing. More thought should be put as to
where this ought to be too, though the device specific script is likely
to be the best spot. The script/add_device.py file needs to be
implemented to allow users to easily add more devices to the configs.